Create chocolate factory
A chocolate factory is packing chocolates into the packets. The chocolate packets here represent an array arrt of N number of integer values. The task is to find the empty packets(0) of chocolate and push it to the end of the conveyor belt(array).

For Example: N=7 and arr = [4,5,0,1.9,0,5,0]. There are 3 empty packets in the given set. These 3 empty packets represented as O should be pushed towards the end of the array

Input Format

Read size of the array Read elements of the array line by line

Constraints

1<=n<=10^7

Output Format

Print the elements

Sample Input 0

7
4
5
0
1
9
0
5 
Sample Output 0

4 5 1 9 5 0 0
Sample Input 1

6
6
0
1
8
0
2 
Sample Output 1

6 1 8 2 0 0